+++ src/CModel.h
@@ -53,11 +53,11 @@ protected:
 
 public:
 	// Set the parameters in this model, scaling from a uniform hypercube to physical units as necessary.
-	void GetFreeParameters(float * params, int n_params, bool scale_params);
+	void GetFreeParameters(double * params, int n_params, bool scale_params);
 	vector<string> GetFreeParameterNames();
-	vector< pair<float, float> > GetFreeParamMinMaxes();
-	void SetFreeParameters(float * params, int n_params, bool scale_params);
-	void GetAllParameters(float * params, int n_params);
+	vector< pair<double, double> > GetFreeParamMinMaxes();
+	void SetFreeParameters(double * params, int n_params, bool scale_params);
+	void GetAllParameters(double * params, int n_params);
 
 public:
 	CModel(int n_params);
